requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-genai
sentence-transformers
faiss-cpu
numpy
pymupdf
logger
